requirements.txt
# coding: utf-8
Flask
google-generativeai
python-dotenv
pandas
openpyxl
xlsxwriter
python-docx
markdown